+++ lib/ueberauth_tumblr.ex
@@ -1,0 +1,3 @@
+defmodule UeberauthTumblr do
+  @moduledoc false
+end